requirements.txt
pandas
numpy
matplotlib
scikit-learn
seaborn
sentence-transformers
torch
transformers
requests
spacy
collections-extended
newspaper3k
lxml_html_clean
jupyter
python-dotenv
